prototype/src/checkout/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=checkout/requirements.txt checkout/requirements.in
#
grpcio==1.33.2            # via -r checkout/requirements.in
protobuf==3.13.0          # via -r checkout/requirements.in
six==1.15.0               # via grpcio, protobuf
sqlitedict==1.7.0         # via -r checkout/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
